requirements.txt
streamlit>=1.32.0
langchain-core>=0.2.0
langchain-community>=0.2.0        # ← needed for DirectoryLoader

langchain-openai>=0.1.0            # if you call OpenAI tools through LangChain
Office365-REST-Python-Client>=2.5.0  # SharePoint integration

# Configuration and validation
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0
python-json-logger>=2.0.0i>=1.13.3
msal>=1.25.0  # for Outlook authentication
requests>=2.31
langchain>=0.1.9
faiss-cpu
tiktoken
python-dotenv
pydub
openai-whisper

# --- voice & media ---
streamlit-webrtc>=0.46           # microphone / WebRTC bridge
soundfile>=0.12                  # write WAV bytes
av>=12.0.0                       # required by streamlit‑webrtc

# --- numerical utils ---
numpy>=1.26

langchain-core>=0.2.0
langchain-community>=0.2.0        # ← needed for DirectoryLoader

langchain-openai>=0.1.0            # if you call OpenAI tools through LangChain
Office365-REST-Python-Client>=2.5.0  # SharePoint integration




